pyproject.toml
[tool.poetry]
name = "repl_python3_Scyllare"
version = "0.1.0"
description = ""
authors = ["Karonpa <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.0"
discord = "^1.7.3"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
